+++ spine-flutter/CHANGELOG.md
@@ -1,5 +1,10 @@
+# 0.0.4
+* Clean-up `fromAsset()` factory methods so the atlas comes before skeleton data file name.
+* Rename `Vector2` to `Vec2`.
+* Make the bundle configurable in `SpineWidget.asset()`.
+
 # 0.0.3
-* Lower macOS deployment target to 10.11
+* Lower macOS deployment target to 10.11.
 
 # 0.0.2
 * Fix package name in build system `spine_flutter` > `esotericsoftware_spine_flutter`.